+++ CHANGELOG.md
@@ -5,4 +5,4 @@ Listen and send keyboard and mouse events on Windows, Linux and MacOS. Forked fr
 ## [1.0.0] - 2024-08-22
 
 - Forked from `rdev`, updated almost a year ago.
-- People have been asking for a way to cancel. `cancel_flag: Arc<AtomicBool>` added on Linux `grab`.
+- Add `cancel_rx: Receiver<()>` to allow cancel listeneing.